+++ include/yaclib/async/task.hpp
@@ -2,8 +2,8 @@
 
 #include <yaclib/async/detail/result_core.hpp>
 #include <yaclib/async/future.hpp>
-#include <yaclib/executor/executor.hpp>
-#include <yaclib/executor/inline.hpp>
+#include <yaclib/exe/executor.hpp>
+#include <yaclib/exe/inline.hpp>
 #include <yaclib/util/helper.hpp>
 #include <yaclib/util/type_traits.hpp>
