+++ include/yaclib/coro/detail/await_awaiter.hpp
@@ -2,7 +2,7 @@
 
 #include <yaclib/async/future.hpp>
 #include <yaclib/coro/coroutine.hpp>
-#include <yaclib/executor/thread_pool.hpp>
+#include <yaclib/exe/thread_pool.hpp>
 #include <yaclib/util/detail/atomic_counter.hpp>
 #include <yaclib/util/type_traits.hpp>
