+++ test/example/async/when_any.cpp
@@ -4,7 +4,7 @@
  */
 #include <yaclib/algo/when_any.hpp>
 #include <yaclib/async/run.hpp>
-#include <yaclib/executor/thread_pool.hpp>
+#include <yaclib/exe/thread_pool.hpp>
 #include <yaclib/util/intrusive_ptr.hpp>
 #include <yaclib/util/result.hpp>
