+++ test/unit/async/task.cpp
@@ -1,5 +1,5 @@
 #include <yaclib/async/task.hpp>
-#include <yaclib/executor/thread_pool.hpp>
+#include <yaclib/exe/thread_pool.hpp>
 #include <yaclib/util/result.hpp>
 
 #include <thread>
