+++ test/unit/coro/async_mutex_stress.cpp
@@ -5,7 +5,7 @@
 #include <yaclib/coro/await.hpp>
 #include <yaclib/coro/future_traits.hpp>
 #include <yaclib/coro/on.hpp>
-#include <yaclib/executor/thread_pool.hpp>
+#include <yaclib/exe/thread_pool.hpp>
 
 #include <limits>
 #include <vector>
