+++ test/unit/coro/on.cpp
@@ -4,8 +4,8 @@
 #include <yaclib/coro/await.hpp>
 #include <yaclib/coro/future_traits.hpp>
 #include <yaclib/coro/on.hpp>
-#include <yaclib/executor/strand.hpp>
-#include <yaclib/executor/thread_pool.hpp>
+#include <yaclib/exe/strand.hpp>
+#include <yaclib/exe/thread_pool.hpp>
 
 #include <array>
 #include <exception>
