+++ test/unit/exe/strand.cpp
@@ -1,9 +1,9 @@
 #include <util/time.hpp>
 
-#include <yaclib/executor/executor.hpp>
-#include <yaclib/executor/strand.hpp>
-#include <yaclib/executor/submit.hpp>
-#include <yaclib/executor/thread_pool.hpp>
+#include <yaclib/exe/executor.hpp>
+#include <yaclib/exe/strand.hpp>
+#include <yaclib/exe/submit.hpp>
+#include <yaclib/exe/thread_pool.hpp>
 #include <yaclib/util/detail/node.hpp>
 #include <yaclib/util/intrusive_ptr.hpp>
